fixed timestamp in BCUpdate to show localtime, not gmt

+++ script/daily_BCUpdate.pl
@@ -5,11 +5,19 @@ use FindBin '$Bin';
 use lib "$Bin/../lib";
 use CovidSchools::BCUpdate;
 use File::Temp 'tempdir';
+use Fcntl;
+use DB_File;
 
 use constant CACHE_DIR=> '/tmp/BCUpdateCache';
 use constant GDRIVE   => '/BC';
 
 mkdir CACHE_DIR unless -e CACHE_DIR;
+
+# open/create DB_File to cache GDrive path to URL info
+my %GDriveCache;
+tie %GDriveCache,'DB_File',CACHE_DIR.'/GDriveCache.db',O_CREAT|O_RDWR,0666,$DB_HASH
+    or die "Cannot open dbfile: $!";
+
 my $bc = CovidSchools::BCUpdate->new(CLEAN_CSV => "$ENV{HOME}/Dropbox/BC_Automation/daily_update",
 				     'CACHE_DIR' => CACHE_DIR,
     );
@@ -32,19 +40,35 @@ print STDERR "Writing out copy of parsed tracker file\n";
 $bc->write_tracker_file($table);
 
 print STDERR "Mirroring documentation\n";
-$bc->mirror_articles($table,$dest);
+$bc->mirror_articles($table,$dest,\%GDriveCache,\&make_gdrive_url);
 
 print STDERR "Writing out clean CSV file\n";
 $bc->write_clean_file($table);
 
 exit 0;
 
+sub make_gdrive_url {
+    my $gdrive_path = shift;
+    return $GDriveCache{$gdrive_path} if $GDriveCache{$gdrive_path};
+    my $url = `rclone --drive-shared-with-me link "gdrive:$gdrive_path"`;
+    chomp $url;
+    $GDriveCache{$gdrive_path} = $url if $url;
+    return $url;
+}
+
 1;
+
+
 
 END {
     system "sync";
     sleep 1;
     system "sync";
-    system "fusermount -u $dest";
-    rmdir $dest;
+    if ($dest) {
+	system "fusermount -u $dest";
+	rmdir $dest;
+    }
+    if (%GDriveCache) {
+	untie %GDriveCache;
+    }
 }
